fix: 事务操作-收藏/取消收藏

+++ internal/app/mp2c/controller/api/types.go
@@ -15,7 +15,7 @@ type GetTopicPageListForm struct {
 	controller.PageFrom
 }
 type ChangeTopicLikeForm struct {
-	TopicId int `json:"topicId" form:"topicId" binding:"required" alias:"文章id"`
+	TopicId int64 `json:"topicId" form:"topicId" binding:"required" alias:"文章id"`
 }
 type GetTagForm struct {
 	ID int64 `json:"id" form:"id" binding:"gte=0" alias:"tag id"`
